pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kantu"
version = "1.0.0"
description = "Git-like image gallery management with delta encoding for storage optimization"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.9"
authors = [
    {name = "KanTu Developers", email = "kantu@example.com"}
]
keywords = ["image", "gallery", "delta", "deduplication", "version-control"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "Pillow>=10.0.0",
    "imagehash>=4.3.1",
    "scikit-image>=0.21.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyvips>=2.2.0",
    "flask>=3.0.0",
    "PySide6>=6.5.0",
    "pyqtgraph>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
kantu = "kantu.cli:main"

[project.gui-scripts]
kantu-gui = "kantu.gui:main"

[project.urls]
Homepage = "https://github.com/cycleuser/KanTu"
Documentation = "https://github.com/cycleuser/KanTu#readme"
Repository = "https://github.com/cycleuser/KanTu"
Issues = "https://github.com/cycleuser/KanTu/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["kantu*"]

[tool.setuptools.package-data]
kantu = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"